requirements.txt
markdown>=3.6
requests>=2.31
pyyaml>=6.0.1

